fold bends, drop BendBase
Removed the BendBase class; SBend and RBend now derive directly from ElementBase,
each holding its own copy of the bend field state and apply()/computeFieldHost logic
(duplication accepted). Their edge transforms come from the unified Geometry via the
ElementBase default, so the Euclid3D toCoordinateSystemTrafo bridge is gone. Added
getBendAngle/getEntranceAngle/getChordLength/getDesignPath as ElementBase virtuals
(straight defaults) and removed the three dynamic_cast<BendBase*> sites in OpalBeamline.
MultipoleT marks its bend-angle accessors override.
