+++ src/AbsBeamline/CMakeLists.txt
@@ -7,7 +7,6 @@ add_subdirectory(EndFieldModel)
 set(_SRCS
     AttributeSet.cpp
     BeamlineVisitor.cpp
-    BendBase.cpp
     ConstantEFieldCavity.cpp
     Drift.cpp
     ElementBase.cpp
@@ -34,7 +33,6 @@ set(_SRCS
 set(_HDRS
     AttributeSet.h
     BeamlineVisitor.h
-    BendBase.h
     ConstantEFieldCavity.h
     Drift.h
     ElementBase.h
